Adjust UIAutomation and add global usings

+++ src/Rocketcress.UIAutomation/UIAutomationTestContext.cs
@@ -1,9 +1,8 @@
-﻿using Rocketcress.Core.Base;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
+﻿using Rocketcress.Core;
+using Rocketcress.Core.Base;
+using Rocketcress.UIAutomation.Common;
+using Rocketcress.UIAutomation.Interaction;
 #if !SLIM
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 #endif
 
 namespace Rocketcress.UIAutomation
@@ -14,18 +13,9 @@ namespace Rocketcress.UIAutomation
     public class UIAutomationTestContext : WindowsTestContextBase
     {
         /// <summary>
-        /// Gets the current instance of the <see cref="UIAutomationTestContext"/>.
+        /// Gets the current test settings.
         /// </summary>
-        public static new UIAutomationTestContext CurrentContext { get; private set; }
-
-        /// <summary>
-        /// Gets or sets the current test settings.
-        /// </summary>
-        public new Settings Settings
-        {
-            get => (Settings)base.Settings;
-            set => base.Settings = value;
-        }
+        public new Settings Settings => (Settings)base.Settings;
 
         /// <summary>
         /// Gets or sets the <see cref="Application"/> that is currently under test and is actively tested currently.
@@ -33,22 +23,39 @@ namespace Rocketcress.UIAutomation
         public Application ActiveApplication { get; set; }
 
         /// <summary>
-        /// Gets or sets all <see cref="Application"/>s that are currently under test.
+        /// Gets all <see cref="Application"/>s that are currently under test.
         /// </summary>
-        public List<Application> Applications { get; set; }
+        public List<Application> Applications { get; }
 
+#pragma warning disable CS1572 // XML comment has a param tag, but there is no parameter by that name
+#pragma warning disable SA1612 // Element parameter documentation should match element parameters
         /// <summary>
         /// Initializes a new instance of the <see cref="UIAutomationTestContext"/> class.
         /// </summary>
-        protected UIAutomationTestContext()
+        /// <param name="testContext">The current MSTest test context.</param>
+        /// <param name="settings">The test settings.</param>
+        protected UIAutomationTestContext(
+#if !SLIM
+            TestContext testContext,
+#endif
+            Settings settings)
+            : base(testContext, settings)
         {
+            Applications = new List<Application>();
         }
+#pragma warning restore CS1572 // XML comment has a param tag, but there is no parameter by that name
+#pragma warning restore SA1612 // Element parameter documentation should match element parameters
 
-        /// <inheritdoc />
-        protected override void OnContextCreated(TestContextBase lastContext)
+        protected override void OnInitialize()
         {
-            base.OnContextCreated(lastContext);
-            CurrentContext = this;
+            base.OnInitialize();
+
+            Mouse.IsWaitForControlReadyEnabled = true;
+            ControlUtility.EnsureControlRegistryIsFilled();
+
+            // Set process DPI Aware, so Clicks are working on High DPI screens
+            if (Environment.OSVersion.Version.Major >= 6)
+                WindowsApiHelper.SetProcessDPIAware();
         }
 
         /// <inheritdoc />
@@ -56,76 +63,16 @@ namespace Rocketcress.UIAutomation
         {
             foreach (var app in Applications)
             {
-                if (!app.Process.HasExited)
+                if (app.StartType == ApplicationStartType.Launched && !app.Process.HasExited)
                     app.Process.Kill();
             }
 
-            CurrentContext = null;
+            if (disposing)
+            {
+                Applications.Clear();
+            }
+
             base.Dispose(disposing);
         }
-
-#pragma warning disable CS1572 // XML comment has a param tag, but there is no parameter by that name
-#pragma warning disable SA1612 // Element parameter documentation should match element parameters
-        /// <summary>
-        /// Creates a new <see cref="UIAutomationTestContext"/> as uses it as the current test context. Please make sure to dispose any preexisting <see cref="TestContextBase"/> instances beforehand.
-        /// </summary>
-        /// <typeparam name="T">The type of the context.</typeparam>
-        /// <param name="activationFunc">A function that creates an instance of the wanted test context class.</param>
-        /// <param name="settings">The settings to use during the test.</param>
-        /// <param name="testContext">The MSTest Test Context.</param>
-        /// <param name="initAction">An action that is executed before the new context is set as current context. Add additional information to the object here if needed.</param>
-        /// <returns>The created context.</returns>
-        protected static T CreateContext<T>(
-            Func<T> activationFunc,
-            Settings settings,
-#if !SLIM
-            TestContext testContext,
-#endif
-            Action<T> initAction)
-            where T : UIAutomationTestContext
-        {
-            return TestContextBase.CreateContext<T>(
-                activationFunc,
-                settings,
-#if !SLIM
-                testContext,
-#endif
-                Initialize);
-
-            void Initialize(T ctx)
-            {
-                ctx.Applications = new List<Application>();
-                initAction?.Invoke(ctx);
-            }
-        }
-
-        /// <summary>
-        /// Creates a new <see cref="UIAutomationTestContext"/> as uses it as the current test context. Please make sure to dispose any preexisting <see cref="TestContextBase"/> instances beforehand.
-        /// </summary>
-        /// <param name="settings">The settings to use during the test.</param>
-        /// <param name="testContext">The MSTest Test Context.</param>
-        /// <returns>The created context.</returns>
-        [SuppressMessage("StyleCop.CSharp.SpacingRules", "SA1009:Closing parenthesis should be spaced correctly", Justification = "SLIM check")]
-        [SuppressMessage("StyleCop.CSharp.ReadabilityRules", "SA1111:Closing parenthesis should be on line of last parameter", Justification = "SLIM check")]
-        [SuppressMessage("StyleCop.CSharp.ReadabilityRules", "SA1115:Parameter should follow comma", Justification = "SLIM check")]
-        [SuppressMessage("StyleCop.CSharp.SpacingRules", "SA1001:Commas should be spaced correctly", Justification = "SLIM check")]
-        [SuppressMessage("StyleCop.CSharp.ReadabilityRules", "SA1113:Comma should be on the same line as previous parameter", Justification = "SLIM check")]
-        public static UIAutomationTestContext CreateContext(
-            Settings settings
-#if !SLIM
-          , TestContext testContext
-#endif
-            )
-        {
-            return CreateContext(
-                () => new UIAutomationTestContext(),
-                settings,
-#if !SLIM
-                testContext,
-#endif
-                null);
-        }
-#pragma warning restore SA1612 // Element parameter documentation should match element parameters
-#pragma warning restore CS1572 // XML comment has a param tag, but there is no parameter by that name
     }
 }
